Change error handling to user error package for better error message generation

+++ domain/user.go
@@ -5,6 +5,6 @@ type User struct {
 	Email    string    `db:"email" json:"email" validate:"required,email"`
 	Password string    `db:"password" json:"password" validate:"required"`
 	ID       int64     `db:"id"`
-	Credit   float64   `db:"credit"`
+	Credit   int64     `db:"credit"`
 	Roles    *[]string `json:"roles"`
 }
